@@ -44,98 +41,98 @@
   "bugs": {
     "url": "https://github.com/Brickies/vui/issues"
   },
-  "dependencies": {
-    "raf.js": "0.0.4"
-  },
   "peerDependencies": {
     "vue": "^2.5.2"
   },
+  "dependencies": {
+    "vue": "^2.5.2",
+    "vue-router": "^3.0.1"
+  },
   "devDependencies": {
-    "autoprefixer": "^6.7.5",
-    "avoriaz": "^1.11.1",
-    "babel-cli": "^6.14.0",
-    "babel-core": "^6.17.0",
-    "babel-eslint": "^6.1.2",
-    "babel-helper-vue-jsx-merge-props": "^2.0.2",
-    "babel-loader": "^6.2.5",
-    "babel-plugin-module-resolver": "^2.2.0",
+    "autoprefixer": "^7.1.2",
+    "avoriaz": "^6.3.0",
+    "babel-core": "^6.22.1",
+    "babel-eslint": "^8.2.1",
+    "babel-helper-vue-jsx-merge-props": "^2.0.3",
+    "babel-loader": "^7.1.1",
+    "babel-plugin-istanbul": "^4.1.1",
     "babel-plugin-syntax-jsx": "^6.18.0",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-plugin-transform-vue-jsx": "^3.3.0",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.16.0",
-    "babel-runtime": "^6.11.0",
-    "chai": "^3.5.0",
-    "chalk": "^1.1.3",
-    "cheerio": "^0.22.0",
+    "babel-plugin-transform-runtime": "^6.22.0",
+    "babel-plugin-transform-vue-jsx": "^3.5.0",
+    "babel-preset-env": "^1.3.2",
+    "babel-preset-stage-2": "^6.22.0",
+    "chai": "^4.1.2",
+    "chalk": "^2.0.1",
     "child_process": "^1.0.2",
     "copy-webpack-plugin": "^4.0.1",
-    "cp-cli": "^1.0.2",
-    "cross-env": "^3.1.3",
-    "css-loader": "^0.24.0",
-    "decamelize": "^1.2.0",
+    "cp-cli": "^1.1.0",
+    "cross-env": "^5.0.1",
+    "css-loader": "^0.28.0",
+    "eslint": "^4.15.0",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-friendly-formatter": "^3.0.0",
     "eslint-loader": "^1.7.1",
-    "eslint-plugin-vue": "^2.0.1",
-    "extract-text-webpack-plugin": "^2.0.0-beta.5",
-    "felint": "^0.5.0-alpha.3",
-    "file-loader": "^0.9.0",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-node": "^5.2.0",
+    "eslint-plugin-promise": "^3.4.0",
+    "eslint-plugin-standard": "^3.0.1",
+    "eslint-plugin-vue": "^4.0.0",
+    "extract-text-webpack-plugin": "^3.0.0",
+    "fast-vue-md-loader": "^1.0.3",
+    "file-loader": "^1.1.4",
     "file-save": "^0.2.0",
-    "gh-pages": "^0.12.0",
-    "gulp": "^3.9.1",
-    "gulp-util": "^3.0.8",
-    "highlight.js": "^9.8.0",
-    "html-loader": "^0.4.3",
-    "html-webpack-plugin": "^2.22.0",
-    "inject-loader": "^3.0.0-beta2",
-    "inquirer": "^4.0.1",
-    "isparta-loader": "^2.0.0",
-    "json-loader": "^0.5.4",
-    "json-templater": "^1.0.4",
-    "karma": "^1.5.0",
-    "karma-chrome-launcher": "^2.0.0",
+    "friendly-errors-webpack-plugin": "^1.6.1",
+    "gh-pages": "^1.1.0",
+    "highlight.js": "^9.12.0",
+    "html-webpack-plugin": "^2.30.1",
+    "inject-loader": "^3.0.0",
+    "inquirer": "^5.1.0",
+    "karma": "^1.4.1",
     "karma-coverage": "^1.1.1",
     "karma-mocha": "^1.3.0",
-    "karma-phantomjs-launcher": "^1.0.4",
-    "karma-sinon-chai": "^1.2.4",
+    "karma-phantomjs-launcher": "^1.0.2",
+    "karma-phantomjs-shim": "^1.4.0",
+    "karma-sinon-chai": "^1.3.1",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-spec-reporter": "0.0.30",
+    "karma-spec-reporter": "0.0.31",
     "karma-webpack": "^2.0.2",
-    "lerna": "2.0.0-beta.31",
-    "lolex": "^1.5.1",
-    "markdown-it": "^6.1.1",
-    "markdown-it-anchor": "^2.5.0",
-    "markdown-it-container": "^2.0.0",
     "mocha": "^3.2.0",
-    "optimize-css-assets-webpack-plugin": "^1.3.0",
-    "postcss": "^5.1.2",
-    "postcss-easy-import": "^2.0.0",
-    "postcss-loader": "^1.3.3",
-    "precss": "^1.4.0",
-    "prismjs": "^1.5.1",
-    "progress-bar-webpack-plugin": "^1.9.3",
-    "rimraf": "^2.5.4",
-    "run-sequence": "^1.2.2",
-    "saladcss-bem": "^0.0.1",
-    "sinon": "^1.17.7",
+    "node-notifier": "^5.1.2",
+    "optimize-css-assets-webpack-plugin": "^3.2.0",
+    "ora": "^1.2.0",
+    "phantomjs-prebuilt": "^2.1.14",
+    "portfinder": "^1.0.13",
+    "postcss-import": "^11.0.0",
+    "postcss-loader": "^2.0.8",
+    "postcss-mixins": "^6.2.0",
+    "postcss-salad": "^2.0.1",
+    "postcss-url": "^7.2.1",
+    "precss": "^3.1.2",
+    "rimraf": "^2.6.0",
+    "saladcss-bem": "^0.0.2",
+    "semver": "^5.3.0",
+    "shelljs": "^0.7.6",
+    "sinon": "^4.0.0",
     "sinon-chai": "^2.8.0",
-    "style-loader": "^0.13.1",
-    "theaterjs": "^3.0.0",
-    "transliteration": "^1.1.11",
-    "uppercamelcase": "^1.1.0",
-    "url-loader": "^0.5.7",
-    "vue": "^2.5.2",
-    "vue-hot-reload-api": "^1.3.3",
-    "vue-html-loader": "^1.2.3",
+    "uglifyjs-webpack-plugin": "^1.1.1",
+    "uppercamelcase": "^3.0.0",
+    "url-loader": "^0.5.8",
     "vue-loader": "^13.3.0",
-    "vue-markdown-loader": "^0.6.1",
-    "vue-router": "^3.0.1",
-    "vue-style-loader": "^1.0.0",
+    "vue-markdown-loader": "^2.4.0",
+    "vue-style-loader": "^3.0.1",
     "vue-template-compiler": "^2.5.2",
-    "vue-template-es2015-compiler": "^1.4.2",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^1.16.3",
-    "webpack-merge": "^2.0.0",
-    "webpack-node-externals": "^1.5.4",
-    "webpack-vendor-chunk-plugin": "^1.0.0"
-  }
+    "webpack": "^3.6.0",
+    "webpack-bundle-analyzer": "^2.9.0",
+    "webpack-dev-server": "^2.9.1",
+    "webpack-merge": "^4.1.0"
+  },
+  "engines": {
+    "node": ">= 6.0.0",
+    "npm": ">= 3.0.0"
+  },
+  "browserslist": [
+    "> 1%",
+    "last 2 versions",
+    "not ie <= 8"
+  ]
 }
