@@ -1,12 +1,4 @@
 <style scoped>
-button {
-  padding: 5px 10px;
-  background: #fff;
-  border: 1px solid #eee;
-  border-radius: 5px;
-  color: #555;
-  font-size: 14px;
-}
 .modal-text {
   text-align: center;
   margin-top: 5px;
@@ -16,23 +8,22 @@ button {
   <section class="demo-toast">
     <h1 class="demo-title">Dialog 弹框</h1>
     <example-block title="基础用法">
-      <button @click="showSimpleDialog">普通 Dialog</button>
+      <v-button @click="showSimpleDialog">普通 Dialog</v-button>
     </example-block>
     <example-block title="自定义 HTML">
-      <button @click="showHtmlDialog">HTML Dialog</button>
+      <v-button @click="showHtmlDialog">HTML Dialog</v-button>
     </example-block>
     <example-block title="Dialog 模板">
-      <button @click="showDialogTpl">Dialog Template</button>
+      <v-button @click="showDialogTpl">Dialog Template</v-button>
     </example-block>
     <v-dialog
-    title="Dialog 模板"
-    cancelText="取消"
-    okText="确认"
-    content="测试 Dialog，测试 Dialog，测试 Dialog~~~"
-    :show="showDialog"
-    :onCancel="close"
-    :onOk="close"
-    >
+      title="Dialog 模板"
+      cancelText="取消"
+      okText="确认"
+      content="测试 Dialog，测试 Dialog，测试 Dialog~~~"
+      :show="showDialog"
+      :onCancel="close"
+      :onOk="close">
       <p class="modal-text">Dialog Template slot ！！！</p>
     </v-dialog>
   </section>
