@@ -19,8 +19,6 @@ import MobileNav from './mobile-nav';
 export default {
   data() {
     return {
-      highlights: [],
-      navState: [],
       data: navConfig['zh-CN'],
       base: '/component'
     };
@@ -32,34 +30,34 @@ export default {
 };
 </script>
 
-<style>
-  .side-nav {
-    width: 100%;
-    box-sizing: border-box;
-    padding: 90px 15px 20px;
-    position: relative;
-    z-index: 1;
+<style lang="postcss">
+.side-nav {
+  width: 100%;
+  box-sizing: border-box;
+  padding: 90px 15px 20px;
+  position: relative;
+  z-index: 1;
 
-    .vui-title,
-    .vui-desc {
-      text-align: center;
-      font-weight: normal;
-      user-select: none;
-    }
-
-    .vui-title {
-      padding-top: 40px;
-      height: 0;
-      overflow: hidden;
-      background: url(https://raw.githubusercontent.com/xuqiang521/vui/master/src/assets/logo.png) center center no-repeat;
-      background-size: 40px 40px;
-      margin-bottom: 10px;
-    }
-
-    .vui-desc {
-      font-size: 14px;
-      color: #666;
-      margin-bottom: 50px;
-    }
+  .vui-title,
+  .vui-desc {
+    text-align: center;
+    font-weight: normal;
+    user-select: none;
   }
+
+  .vui-title {
+    padding-top: 40px;
+    height: 0;
+    overflow: hidden;
+    background: url(https://raw.githubusercontent.com/xuqiang521/vui/master/src/assets/logo.png) center center no-repeat;
+    background-size: 40px 40px;
+    margin-bottom: 10px;
+  }
+
+  .vui-desc {
+    font-size: 14px;
+    color: #666;
+    margin-bottom: 50px;
+  }
+}
 </style>
