@@ -1,54 +1,51 @@
-import Vue from 'vue';
-import VueRouter from 'vue-router';
-import App from './DocsApp';
-import navConfig from './nav.config.json';
-import routes from './router.config';
-import Vui from 'src/index.js';
-import isMobile from './isMobile.js';
-import { VERSION } from '../../package.json';
+// The Vue build version to load with the `import` command
+// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
+import Vue from 'vue'
+import VueRouter from 'vue-router'
+import App from './App'
+import registerRoute from './router.config'
+import navConfig from './nav.config'
+import isMobile from './is-mobile'
+import Vui from 'src/index.js'
+import 'packages/vui-css/src/index.css'
+import '../assets/docs.css'
 
-const isProduction = process.env.NODE_ENV === 'production';
-const global = {
-  version: VERSION
-};
-window._global = global;
+Vue.config.productionTip = false
+Vue.use(Vui)
+Vue.use(VueRouter)
 
-import '../assets/docs.css';
-import 'packages/vui-css/src/index.css';
-
-Vue.use(Vui);
-Vue.use(VueRouter);
-
-const routesConfig = routes(navConfig);
+const routesConfig = registerRoute(navConfig)
+const isProduction = process.env.NODE_ENV === 'production'
 routesConfig.push({
   path: '/',
-  redirect: '/component/intro'
-});
+  redirect: '/component/button'
+})
 routesConfig.push({
   path: '/component',
-  redirect: '/component/intro'
-});
+  redirect: '/component/button'
+})
 
 const router = new VueRouter({
-  mode: 'hash',
   base: isProduction ? '/vui/' : __dirname,
   routes: routesConfig
-});
+})
 router.beforeEach((route, redirect, next) => {
   if (route.path !== '/') {
-    window.scrollTo(0, 0);
+    window.scrollTo(0, 0)
   }
 
-  const pathname = isProduction ? '/vui/mobile' : '/mobile.html';
+  const pathname = isProduction ? '/vui/mobile' : '/mobile.html'
   if (isMobile) {
-    window.location.replace(pathname);
-    return;
+    window.location.replace(pathname)
+    return
   }
-  document.title = route.meta.title || document.title;
-  next();
-});
-
-new Vue({ // eslint-disable-line
-  render: h => h(App),
-  router
-}).$mount('#app-container');
+  document.title = route.meta.title || document.title
+  next()
+})
+/* eslint-disable no-new */
+new Vue({
+  el: '#app-container',
+  router,
+  components: { App },
+  template: '<App/>'
+})
