@@ -1,44 +1,45 @@
-import Vue from 'vue';
-import VueRouter from 'vue-router';
-import App from './MobileApp';
-import navConfig from './nav.config.json';
-import routes from './router.config';
-import Vui from 'src/index';
-import isMobile from './isMobile.js';
+import Vue from 'vue'
+import VueRouter from 'vue-router'
+import MobileApp from './MobileApp'
+import registerRoute from './router.config'
+import navConfig from './nav.config'
+import Vui from 'src/index'
+import isMobile from './is-mobile.js'
 
-import 'packages/vui-css/src/index.css';
+import 'packages/vui-css/src/index.css'
 
-import DemoList from './components/demo-list.vue';
+import DemoList from './components/demo-list.vue'
 
-Vue.use(Vui);
-Vue.use(VueRouter);
+Vue.use(Vui)
+Vue.use(VueRouter)
 
-const isProduction = process.env.NODE_ENV === 'production';
-const routesConfig = routes(navConfig, true);
+const isProduction = process.env.NODE_ENV === 'production'
+const routesConfig = registerRoute(navConfig, true)
 routesConfig.push({
   path: '/',
-  component: DemoList.default || DemoList
-});
+  component: DemoList
+})
 const router = new VueRouter({
-  mode: 'hash',
   base: isProduction ? '/vui/' : __dirname,
   routes: routesConfig
-});
-
+})
+console.log(isMobile)
 router.beforeEach((route, redirect, next) => {
   if (route.path !== '/') {
-    window.scrollTo(0, 0);
+    window.scrollTo(0, 0)
   }
-  const pathname = isProduction ? '/vui/' : '/';
+  const pathname = isProduction ? '/vui/' : '/'
   if (!isMobile) {
-    window.location.replace(pathname);
-    return;
+    window.location.replace(pathname)
+    return
   }
-  document.title = route.meta.title || document.title;
-  next();
-});
-
-new Vue({ // eslint-disable-line
-  render: h => h(App),
-  router
-}).$mount('#app-container');
+  document.title = route.meta.title || document.title
+  next()
+})
+/* eslint-disable no-new */
+new Vue({
+  el: '#app-container',
+  router,
+  components: { MobileApp },
+  template: '<MobileApp/>'
+})
