@@ -1,37 +1,37 @@
 const registerRoute = (navConfig, isMobile) => {
-  let route = [];
-  let navs = navConfig['zh-CN'];
+  let route = []
+  let navs = navConfig['zh-CN']
   navs.forEach(nav => {
     if (isMobile && !nav.showInMobile) {
-      return;
+      return
     }
 
     if (nav.groups) {
       nav.groups.forEach(group => {
         group.list.forEach(nav => {
-          addRoute(nav);
-        });
-      });
+          addRoute(nav)
+        })
+      })
     } else if (nav.children) {
       nav.children.forEach(nav => {
-        addRoute(nav);
-      });
+        addRoute(nav)
+      })
     } else {
-      addRoute(nav);
+      addRoute(nav)
     }
-  });
+  })
 
-  function addRoute(page) {
+  function addRoute (page) {
     const component = isMobile
       ? require(`../pages${page.path}.vue`)
-      : require(`../docs${page.path}.md`);
+      : require(`../docs${page.path}.md`)
     route.push({
       path: '/component' + page.path,
       component: component.default || component
-    });
+    })
   }
 
-  return route;
-};
+  return route
+}
 
-export default registerRoute;
+export default registerRoute
