@@ -1,13 +1,13 @@
-var gulp = require('gulp');
-var postcss = require('gulp-postcss');
-var cssmin = require('gulp-cssmin');
-var salad = require('postcss-salad')(require('./salad.config.json'));
+const gulp = require('gulp')
+const postcss = require('gulp-postcss')
+const cssmin = require('gulp-cssmin')
+const salad = require('postcss-salad')(require('../../salad.config.json'))
 
-gulp.task('compile', function() {
+gulp.task('compile', function () {
   return gulp.src('./src/*.css')
     .pipe(postcss([salad]))
     .pipe(cssmin())
-    .pipe(gulp.dest('./lib'));
-});
+    .pipe(gulp.dest('./lib'))
+})
 
-gulp.task('build', ['compile']);
+gulp.task('build', ['compile'])
