@@ -1,11 +1,13 @@
-require('packages/vui-css/src/index.css');
+import Vue from 'vue'
+
+Vue.config.productionTip = false
 
 // require all test files (files that ends with .spec.js)
-const testsReq = require.context('./specs', true, /\.spec$/);
-testsReq.keys().forEach(testsReq);
+const testsContext = require.context('./specs', true, /\.spec$/)
+testsContext.keys().forEach(testsContext)
 
 // require all src files except main.js for coverage.
 // you can also change this to match only the subset of files that
 // you want coverage for.
-const srcReq = require.context('../../src', true, /^\.\/(?!main(\.js)?$)/);
-srcReq.keys().forEach(srcReq);
+const srcContext = require.context('../../src', true, /^\.\/(?!main(\.js)?$)/)
+srcContext.keys().forEach(srcContext)
