@@ -1,32 +1,41 @@
-require('babel-polyfill');
+// This is a karma config file. For more details see
+//   http://karma-runner.github.io/0.13/config/configuration-file.html
+// we are also using it with karma-webpack
+//   https://github.com/webpack/karma-webpack
 
-require('babel-core/register')({
-  presets: [require('babel-preset-es2015')]
-});
+var webpackConfig = require('../../build/webpack.test.conf')
 
-var webpackConfig = require('./get-webpack-conf');
-var travis = process.env.TRAVIS;
-
-module.exports = function(config) {
+module.exports = function karmaConfig (config) {
   config.set({
-    browsers: travis ? ['PhantomJS'] : ['Chrome'],
-    frameworks: ['mocha', 'sinon-chai'],
+    // to run in additional browsers:
+    // 1. install corresponding karma launcher
+    //    http://karma-runner.github.io/0.13/config/browsers.html
+    // 2. add it to the `browsers` array below.
+    // 浏览器
+    browsers: ['PhantomJS'],
+    // 测试框架
+    frameworks: ['mocha', 'sinon-chai', 'phantomjs-shim'],
+    // 测试报告
     reporters: ['spec', 'coverage'],
+    // 测试入口文件
     files: ['./index.js'],
+    // 预处理器 karma-webpack
     preprocessors: {
       './index.js': ['webpack', 'sourcemap']
     },
+    // webpack配置
     webpack: webpackConfig,
+    // webpack中间件
     webpackMiddleware: {
       noInfo: true
     },
+    // 测试覆盖率报告
     coverageReporter: {
       dir: './coverage',
       reporters: [
         { type: 'lcov', subdir: '.' },
         { type: 'text-summary' }
       ]
-    },
-    singleRun: false
-  });
-};
+    }
+  })
+}
