@@ -1,24 +1,24 @@
-import Dialog from 'packages/dialog';
-// import { mount } from 'avoriaz';
+import Dialog from 'packages/dialog/dialog'
+// import { mount } from 'avoriaz'
 
 describe('Dialog', () => {
-  let wrapper;
+  let wrapper
   afterEach(() => {
-    wrapper && wrapper.destroy();
-  });
+    wrapper && wrapper.destroy()
+  })
 
   it('create a dialog component', () => {
     Dialog({
       title: 'test',
       okText: '确定'
-    });
-    const dialog = document.querySelector('.v-dialog');
-    expect(dialog).not.to.be.underfined;
+    })
+    const dialog = document.querySelector('.v-dialog')
+    expect(dialog).not.to.be.underfined
 
-    const confirmBtn = document.querySelector('.v-dialog-ok');
-    expect(confirmBtn).not.to.be.underfined;
+    const confirmBtn = document.querySelector('.v-dialog-ok')
+    expect(confirmBtn).not.to.be.underfined
 
-    const content = document.querySelector('.v-dialog-content');
-    expect(content).not.to.be.underfined;
-  });
-});
+    const content = document.querySelector('.v-dialog-content')
+    expect(content).not.to.be.underfined
+  })
+})
