@@ -1,21 +1,21 @@
-import Icon from 'packages/icon';
-import { mount } from 'avoriaz';
+import Icon from 'packages/icon'
+import { mount } from 'avoriaz'
 
 describe('Icon', () => {
-  let wrapper;
+  let wrapper
 
   afterEach(() => {
-    wrapper && wrapper.destroy();
-  });
+    wrapper && wrapper.destroy()
+  })
 
   it('create a address icon', () => {
     wrapper = mount(Icon, {
       propsData: {
         name: 'address'
       }
-    });
+    })
 
-    expect(wrapper.hasClass('v-icon')).to.be.true;
-    expect(wrapper.hasClass('v-icon-address')).to.be.true;
-  });
-});
+    expect(wrapper.hasClass('v-icon')).to.be.true
+    expect(wrapper.hasClass('v-icon-address')).to.be.true
+  })
+})
