@@ -1,21 +1,19 @@
-import Picker from 'packages/picker';
-import { mount } from 'avoriaz';
+import Picker from 'packages/picker'
+import { mount } from 'avoriaz'
 
 describe('Picker', () => {
-  let wrapper;
+  let wrapper
 
   afterEach(() => {
-    wrapper && wrapper.destroy();
-  });
+    wrapper && wrapper.destroy()
+  })
 
   it('create a pick', () => {
     wrapper = mount(Picker, {
       propsData: {
         type: 'date'
       }
-    });
-    expect(wrapper.hasClass('m-datetime')).not.to.be.true;
-    wrapper.simulate('click')
-    expect(wrapper.data().picker).not.be.undefined;
-  });
-});
+    })
+    expect(wrapper.hasClass('m-datetime')).not.to.be.true
+  })
+})
