@@ -1,34 +1,34 @@
-import Switch from 'packages/switch';
-import { mount } from 'avoriaz';
+import Switch from 'packages/switch'
+import { mount } from 'avoriaz'
 
 describe('Switch', () => {
-  let wrapper;
+  let wrapper
 
   afterEach(() => {
-    wrapper && wrapper.destroy();
-  });
+    wrapper && wrapper.destroy()
+  })
 
   it('create on switch', () => {
     wrapper = mount(Switch, {
       propsData: {
         value: true
       }
-    });
+    })
 
-    expect(wrapper.hasClass('v-switch')).to.be.true;
-    expect(wrapper.hasClass('active')).to.be.true;
-  });
+    expect(wrapper.hasClass('v-switch')).to.be.true
+    expect(wrapper.hasClass('active')).to.be.true
+  })
 
   it('create off switch', () => {
     wrapper = mount(Switch, {
       propsData: {
         value: false
       }
-    });
+    })
 
-    expect(wrapper.hasClass('v-switch')).to.be.true;
-    expect(wrapper.hasClass('inactive')).to.be.true;
-  });
+    expect(wrapper.hasClass('v-switch')).to.be.true
+    expect(wrapper.hasClass('inactive')).to.be.true
+  })
 
   it('loading switch should be unclickable', () => {
     wrapper = mount(Switch, {
@@ -36,23 +36,23 @@ describe('Switch', () => {
         loading: true,
         value: true
       }
-    });
+    })
 
-    expect(wrapper.hasClass('active')).to.be.true;
-    wrapper.simulate('click');
-    expect(wrapper.hasClass('active')).to.be.true;
-  });
+    expect(wrapper.hasClass('active')).to.be.true
+    wrapper.trigger('click')
+    expect(wrapper.hasClass('active')).to.be.true
+  })
 
   it('create disabled switch', () => {
     wrapper = mount(Switch, {
       propsData: {
         disabled: true
       }
-    });
+    })
 
-    expect(wrapper.hasClass('v-switch')).to.be.true;
-    expect(wrapper.hasClass('v-switch--disabled')).to.be.true;
-  });
+    expect(wrapper.hasClass('v-switch')).to.be.true
+    expect(wrapper.hasClass('v-switch--disabled')).to.be.true
+  })
 
   it('disabled switch should be unclickable', () => {
     wrapper = mount(Switch, {
@@ -60,10 +60,10 @@ describe('Switch', () => {
         disabled: true,
         value: false
       }
-    });
+    })
 
-    expect(wrapper.hasClass('inactive')).to.be.true;
-    wrapper.simulate('click');
-    expect(wrapper.hasClass('inactive')).to.be.true;
-  });
-});
+    expect(wrapper.hasClass('inactive')).to.be.true
+    wrapper.trigger('click')
+    expect(wrapper.hasClass('inactive')).to.be.true
+  })
+})
