@@ -1,31 +1,30 @@
-import Toast from 'packages/toast';
-// import { mount } from 'avoriaz';
+import Toast from 'packages/toast/toast.js'
+// import { mount } from 'avoriaz'
 
 describe('Toast', () => {
-  let wrapper;
+  let wrapper
   afterEach(() => {
-    wrapper && wrapper.destroy();
-  });
+    wrapper && wrapper.destroy()
+  })
 
   it('create a simple toast', () => {
-    Toast('a toast message');
-    var toast = document.querySelector('.v-toast');
-    expect(toast).not.to.be.underfined;
-
-      setTimeout(() => {
-        expect(toast).to.be.underfined;
-      }, 2001);
-  });
-
-  it('create html toast', () => {
-    Toast('<strong style="font-size: 20px;">HTML 文字提示~</strong>');
-    var toast = document.querySelector('.v-toast');
-
-    expect(toast).not.to.be.underfined;
+    Toast('a toast message')
+    var toast = document.querySelector('.v-toast')
+    expect(toast).not.to.be.underfined
 
     setTimeout(() => {
-      expect(toast).to.be.underfined;
-    }, 2001);
-  });
-});
+      expect(toast).to.be.underfined
+    }, 2001)
+  })
 
+  it('create html toast', () => {
+    Toast('<strong style="font-size: 20px">HTML 文字提示~</strong>')
+    var toast = document.querySelector('.v-toast')
+
+    expect(toast).not.to.be.underfined
+
+    setTimeout(() => {
+      expect(toast).to.be.underfined
+    }, 2001)
+  })
+})
